LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/d-expected.txt


FAIL d responsive to inherited d changes assert_equals: expected "path(\"M 0 0 H 300\")" but got "none"
